images/requirements.txt
certifi==2020.6.20
chardet==3.0.4
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.10
Pillow==7.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.48.0
